requirements.txt
aiohttp==1.0.5
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.0
autopep8==1.4
awscli==1.14.44
beautifulsoup4==4.6.0
boto3==1.9.8
botocore==1.12.8
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
cryptography==2.1.4
cycler==0.10.0
decorator==4.1.2
discord==0.0.2
discord.py==0.16.12
dj-database-url==0.5.0
Django==2.1.1
django-cors-headers==2.4.0
django-heroku==0.3.1
docutils==0.14
gunicorn==19.9.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
isort==4.3.4
jmespath==0.9.3
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
lxml==4.2.1
Markdown==2.6.9
matplotlib==2.1.1
mccabe==0.6.1
multidict==4.4.1
mysqlclient==1.3.13
numpy==1.13.3
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.2.0
psutil==5.4.2
psycopg2==2.7.5
pyasn1==0.4.2
pycairo==1.16.2
pycodestyle==2.4.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.0
python-dateutil==2.7.3
python-debian==0.1.32
python-debianbts==2.7.2
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.1.13
SecretStorage==2.3.1
Send2Trash==1.4.2
simplejson==3.13.2
six==1.11.0
systemd-python==234
typed-ast==1.1.0
urllib3==1.23
virtualenv==16.0.0
wadllib==1.3.2
webencodings==0.5
websockets==3.4
whitenoise==4.1
wrapt==1.10.11
xcffib==0.5.1
zope.interface==4.3.2
